+++ auth.go
@@ -29,13 +29,14 @@ func NewAuth(logger *zap.Logger, memory Memory) *Auth {
 // given scope. If the user is not permitted access this function returns
 // ErrNotAllowed.
 //
-// Scopes are interpreted in a hierarchical way where scope A can be contained
-// in scope B if B is a prefix to A. For example, you can check if a user is
-// allowed to read or write from the "Example" API by checking the
-// "api.example.read" or "api.example.write" scope. When you grant the scope to
-// a user you can now either decide only to grant the very specific
-// "api.example.read" scope which means the user will not have write permissions
-// or you can allow people write-only access via "api.example.write".
+// Scopes are interpreted in a hierarchical way where scope A can contain scope B
+// if A is a prefix to B. For example, you can check if a user is allowed to
+// read or write from the "Example" API by checking the "api.example.read" or
+// "api.example.write" scope. When you grant the scope to a user you can now
+// either decide only to grant the very specific "api.example.read" scope which
+// means the user will not have write permissions or you can allow people
+// write-only access via "api.example.write".
+//
 // Alternatively you can also grant any access to the Example API via "api.example"
 // which includes both the read and write scope beneath it. If you choose to you
 // could also allow even more general access to everything in the api via the
@@ -81,11 +82,16 @@ func (a *Auth) loadPermissions(key string) ([]string, error) {
 	return permissions, nil
 }
 
-// Grant adds a new permission scope to the given user. When a scope was granted
-// to a specific user it can be checked later via CheckPermission(…). The empty
-// scope cannot be granted and trying to do so will result in an error. If you
-// want to grant access to all scopes you should prefix them with a common scope
-// such as "root." or "api.".
+// Grant adds a permission scope to the given user. When a scope was granted
+// to a specific user it can be checked later via CheckPermission(…).
+//
+// Note that granting a scope is an idempotent operations so granting the same
+// scope multiple times is a safe operation and will not change the internal
+// permissions that are written to the Memory.
+//
+// The empty scope cannot be granted and trying to do so will result in an error.
+// If you want to grant access to all scopes you should prefix them with a
+// common scope such as "root." or "api.".
 func (a *Auth) Grant(scope, userID string) error {
 	if scope == "" {
 		return errors.New("scope cannot be empty")
@@ -97,7 +103,17 @@ func (a *Auth) Grant(scope, userID string) error {
 		return errors.WithStack(err)
 	}
 
-	permissions = append(permissions, scope)
+	newScope := true // until proven otherwise
+	for _, p := range permissions {
+		if p == scope {
+			newScope = false
+		}
+	}
+
+	if newScope {
+		permissions = append(permissions, scope)
+	}
+
 	data, err := json.Marshal(permissions)
 	if err != nil {
 		return errors.Wrap(err, "failed to encode permissions as JSON")