+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 **THIS SOFTWARE IS STILL IN ALPHA AND THERE ARE NO GUARANTEES REGARDING API STABILITY YET.**
 
 ## [Unreleased]
-Nothing so far
+- Make Auth.Grant(…) idempotent
 
 ## [v0.7.0] - 2019-04-18
 - Add ReceiveMessageEvent.Data field to allow using the underlying message type of the adapters